Edit payments methods
