+++ src/features/client/components/CustomerAccountPayment/CustomerAccountPayment.scss
@@ -1,9 +1,5 @@
 .customer-payment {
-  &__nav {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: flex-end;
-    gap: 10px;
-    margin-bottom: 20px;
-  }
+  width: 600px;
+  max-width: 100%;
+  margin: 0 auto;
 }
