--- src/features/client/components/CustomerAccountPaymentList/CustomerAccountPaymentList.scss
@@ -1,76 +1,0 @@
-@import './src/constants/variables';
-
-.customer {
-  &-prefer {
-    width: 600px;
-    max-width: 100%;
-    margin: 0 auto 80px;
-
-    @media screen and (max-width: $screen-md) {
-      margin-bottom: 40px;
-    }
-
-    &__title {
-      margin-bottom: 30px;
-      font-weight: 700;
-      font-size: 24px;
-      line-height: 30px;
-      text-align: center;
-
-      @media screen and (max-width: $screen-md) {
-        margin-bottom: 20px;
-        font-size: 18px;
-        line-height: 24px;
-      }
-    }
-
-    &__body {
-      display: flex;
-      justify-content: center;
-      gap: 20px;
-    }
-
-    &__option {
-      font-weight: 700;
-      font-size: 18px;
-      line-height: 24px;
-      color: var(--chakra-colors-gray-400);
-      transition: color 0.3s;
-
-      &.is-active {
-        color: var(--chakra-colors-blue-500);
-      }
-
-      @media screen and (max-width: $screen-md) {
-        font-size: 16px;
-        line-height: 20px;
-      }
-    }
-  }
-
-  &-cardlist {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    width: 600px;
-    max-width: 100%;
-    margin: 0 auto;
-
-    .cardlist {
-      width: 100%;
-
-      &__empty {
-        font-weight: 700;
-        font-size: 20px;
-        line-height: 24px;
-        text-align: center;
-      }
-    }
-
-    &__add {
-      width: 240px;
-      max-width: 100%;
-      margin-top: 40px;
-    }
-  }
-}
