+++ src/features/client/mocks/customer.js
@@ -9,16 +9,7 @@ const customer = {
   photo:
     'https://images.unsplash.com/photo-1619946794135-5bc917a27793?ixlib=rb-0.3.5&q=80&fm=jpg&crop=faces&fit=crop&h=200&w=200&s=b616b2c5b373a80ffc9636ba24f7a4a9',
   payment: {
-    prefer: 'card',
-    cards: [
-      {
-        id: 1,
-        number: '1234 1234 1234 1234',
-        valid: '12/27',
-        cvc: '111',
-        primary: true,
-      },
-    ],
+    prefer: 'paysera',
   },
   addressBook: [
     {
